tests directory refactored
