+++ tests/test_layers_super_resolution.py
@@ -9,8 +9,10 @@ os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 import tensorflow as tf
 import tensorlayer as tl
 
+from tests.utils import CustomTestCase
 
-class Layer_Super_Resolution_Test(unittest.TestCase):
+
+class Layer_Super_Resolution_Test(CustomTestCase):
 
     @classmethod
     def setUpClass(cls):
