+++ tests/test_reuse_mlp.py
@@ -9,6 +9,8 @@ os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 import tensorflow as tf
 import tensorlayer as tl
 
+from tests.utils import CustomTestCase
+
 
 # define the network
 def mlp(x, is_train=True, reuse=False):
@@ -24,7 +26,7 @@ def mlp(x, is_train=True, reuse=False):
     return network
 
 
-class MLP_Reuse_Test(unittest.TestCase):
+class MLP_Reuse_Test(CustomTestCase):
 
     @classmethod
     def setUpClass(cls):
