+++ tests/utils/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from tests.utils.custom_testcase import *
+from tests.utils.list_py_files import *
+from tests.utils.timeout_utils import *
+
+from tests.utils.custom_layers import *
+from tests.utils.custom_networks import *
