+++ tests/utils/custom_layers/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from tests.utils.custom_layers.basic_layers import *
+from tests.utils.custom_layers.inception_blocks import *
