+++ tests/utils/custom_networks/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from tests.utils.custom_networks.inceptionv4 import *
